requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
psycopg2-binary==2.9.1
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
alpaca-trade-api==2.3.0
python-dotenv==0.19.0